+++ logrus/libBioUtils/clustqtci.h
@@ -5,6 +5,7 @@
 
 #include "fullmatrix.h"
 #include "halfmatrix.h"
+#include "pstdint.h"
 
 namespace libBioUtils {
 
@@ -17,12 +18,13 @@ protected:
 		CDistanceMatrix& );
 	static double GetJackDistance( const float*, const float*, size_t, bool, float*, float*,
 		const IMeasure* );
-	static size_t QualityThresholdAll( const CDataMatrix&, float, size_t,
-		const CDistanceMatrix&, std::vector<size_t>& );
+	static uint16_t QualityThresholdAll( const CDataMatrix&, float, size_t,
+		const CDistanceMatrix&, std::vector<uint16_t>& );
 	static void QualityThresholdLargest( const CDataMatrix&, float, const CDistanceMatrix&,
-		const std::vector<bool>&, std::vector<size_t>& );
+		const std::vector<bool>&, std::vector<uint16_t>& );
 	static void QualityThresholdGene( size_t, const CDataMatrix&, float,
-		const CDistanceMatrix&, std::vector<bool>&, std::vector<float>&, std::vector<size_t>& );
+		const CDistanceMatrix&, std::vector<bool>&, std::vector<float>&,
+		std::vector<uint16_t>& );
 };
 
 }
